Conflicts:
	WebContent/WEB-INF/views/budget/todayBudget.jsp

+++ src/report/model/dao/LinearRegression.java
@@ -5,8 +5,8 @@ public class LinearRegression {
 	
 	private List Xdata;  // X축 데이터
     private List YData;  // Y축 데이터
-    private Double result1;			// X축 평균
-    private Double result2;			// Y축 평균
+    private Float result1;			// X축 평균
+    private Float result2;			// Y축 평균
 
     public LinearRegression (List xdata, List YData) {
         this.Xdata = xdata;
@@ -14,45 +14,45 @@ public class LinearRegression {
     }
 
     // 예측값 도출하는 메서드 
-    public Double predictValue (long inputValue) {
-    	Double X1 = (Double)Xdata.get(0);
-    	Double Y1 = (Double)YData.get(0);
-    	Double Xmean = (Double)getXMean(Xdata);
-    	Double Ymean = (Double)getYMean(YData);
-    	Double lineSlope = getLineSlope(Xmean, Ymean, X1, Y1) ;
-    	Double YIntercept = getYIntercept(Xmean , Ymean , lineSlope);
-    	Double prediction = (lineSlope * inputValue) + YIntercept;
+    public Float predictValue (Float inputValue) {
+        Float X1 = (float)Xdata.get(0);
+        Float Y1 = (float)YData.get(0);
+        Float Xmean = (float)getXMean(Xdata);
+        Float Ymean = (float)getYMean(YData);
+        Float lineSlope = getLineSlope(Xmean, Ymean, X1, Y1) ;
+        Float YIntercept = getYIntercept(Xmean , Ymean , lineSlope);
+        Float prediction = (lineSlope * inputValue) + YIntercept;
         return prediction;
     }
     // y = ax + b : a = 가중치 b = 편향 
     
     // 가중치(그래프 각도) 계산 메서드  : 최소 제곱법 이용 
     // a = (x-x평균)(y-y평균)의 합 / (x-x평균)의 합의 제곱  
-    public Double getLineSlope (Double Xmean, Double Ymean, Double X1, Double Y1) {
-    	Double num1 = X1 - Xmean;
-    	Double num2 = Y1 - Ymean;
-    	Double denom = (X1 - Xmean) * (X1 - Xmean);
+    public Float getLineSlope (Float Xmean, Float Ymean, Float X1, Float Y1) {
+        float num1 = X1 - Xmean;
+        float num2 = Y1 - Ymean;
+        float denom = (X1 - Xmean) * (X1 - Xmean);
         return (num1 * num2) / denom;
     }
 
     // 편향(별도로 더해지는 값) 계산 메서드 
-    public Double getYIntercept (Double Xmean, Double Ymean, Double lineSlope) {
+    public float getYIntercept (Float Xmean, Float Ymean, Float lineSlope) {
         return Ymean - (lineSlope * Xmean);
     }
 
     
-    public Double getXMean (List Xdata) {
-        result1 = 0.0 ;
+    public Float getXMean (List Xdata) {
+        result1 = 0.0f ;
         for (Integer i = 0; i < Xdata.size(); i++) {
-            result1 = result1 + (Double)Xdata.get(i);
+            result1 = result1 + (float)Xdata.get(i);
         }
         return result1;
     }
 
-    public Double getYMean (List Ydata) {
-        result2 = 0.0 ;
+    public Float getYMean (List Ydata) {
+        result2 = 0.0f ;
         for (Integer i = 0; i < Ydata.size(); i++) {
-            result2 = result2 + (Double)Ydata.get(i);
+            result2 = result2 + (float)Ydata.get(i);
         }
         return result2;
     }
